Fixed 1 "bug" and added an autostart feature for godot projects
added an autocmd that will automatically open and close godot lsp server when opening or leaving a godot project for the current working directory.

if there is a project.godot file in the current working directory the server will start automatically. and if that directory is left then the autocmd will run again and check for a project.godot file, and if there isn't one it will close the server.